pyproject.toml
[project]
name = "ark_tools"
version = "0.1"
description = ""
dependencies = [
    "Pillow>=10.2.0",
    "httpx>=0.26.0",
    "ruamel-yaml>=0.18.6",
    "aiofiles>=23.2.1",
    # "flatbuffers>=24.3.25",
    "UnityPy @ git+https://github.com/MetalCrab/UnityPy",
    "natsort>=8.4.0",
]
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool.ruff]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "ASYNC", # flake8-async
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "SIM",   # flake8-simplify
    "T10",   # flake8-debugger
    "T20",   # flake8-print
    "PYI",   # flake8-pyi
    "PT",    # flake8-pytest-style
    "Q",     # flake8-quotes
    "F",     # Pyflakes
    "I",     # isort
    "UP",    # pyupgrade
    # "RUF",   # Ruff-specific rules
]
# ignore = ["E402", "B008","C901", "PT023", "T201", "E501"]
# exclude = ["test"]

